% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_run.R
\name{cor_run}
\alias{cor_run}
\title{my cor reporting func}
\usage{
cor_run(df, vars)
}
\arguments{
\item{df}{A data frame containing variables.}
}
\value{
A printed data frame contains pairwise correlations and descriptions between variables
}
\description{
report correlations as a ordered printed data frame
}
