% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-count_na.R
\name{count_na}
\alias{count_na}
\title{my NAs reporting func}
\usage{
count_na(dat, group_var = NULL)
}
\arguments{
\item{dat}{A data frame to be checked for missing values.}

\item{group_var}{(可选) 分组变量名（字符串），如指定则按该变量分组统计缺失。}
}
\value{
一个数据框，包含变量、类型、缺失数、缺失率，若分组则每组一行。
}
\description{
report NAs as a ordered printed data frame
}
