% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-cross_tb.R
\name{cross_tb}
\alias{cross_tb}
\title{交叉表/分组统计函数}
\usage{
cross_tb(dat, var, by)
}
\arguments{
\item{dat}{数据框}

\item{var}{目标变量名（字符串，长度为1）}

\item{by}{分组变量名（字符串向量，长度>=1）}
}
\value{
一个数据框，分类变量返回频数和比例，连续变量返回均值和中位数
}
\description{
交叉表/分组统计函数
}
