% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnas.R
\name{getnas}
\alias{getnas}
\title{my NAs reporting func}
\usage{
getnas(dat)
}
\arguments{
\item{dat}{A data frame containing the grouping variable, event variable, and time variable.}
}
\value{
A printed data frame combining group statistics (Variable，Type，Missing，MissingRate).
}
\description{
report NAs as a ordered printed data frame
}
