% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-cox_run.R
\name{cox_run}
\alias{cox_run}
\title{常规Cox回归分析函数，本函数支持基本的Cox回归分析和时间依赖（长数据）变量的Cox回归分析。}
\usage{
cox_run(
  data,
  time1 = NULL,
  time2 = NULL,
  timediff = NULL,
  event,
  mainvar,
  covars = NULL,
  extra_vars = NULL
)
}
\arguments{
\item{data}{数据框}

\item{time1}{随访开始时间变量名（字符串）}

\item{time2}{随访结束时间变量名（字符串）}

\item{event}{结局变量名（字符串，0/1）}

\item{mainvar}{关注变量名（字符串）}

\item{covars}{协变量名（字符串向量，可为NULL）}

\item{extra_vars}{额外关注变量名（字符串向量，可为NULL）}
}
\value{
一个列表，包含模型对象和格式化结果表
}
\description{
常规Cox回归分析函数，本函数支持基本的Cox回归分析和时间依赖（长数据）变量的Cox回归分析。
}
