% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_run.R
\name{cox_run}
\alias{cox_run}
\title{my Cox Model tools and Group Statistics}
\usage{
cox_run(data, time, event, riskvar, covars, groupvar = NULL, ref = NULL)
}
\arguments{
\item{data}{A data frame containing the grouping variable, event variable, and time variable.}

\item{time}{A string specifying the time variable in \code{df} (in days).}

\item{event}{A string specifying the event variable in \code{df} (1 = event, 0 = censored).}

\item{riskvar}{A string specifying the focused risk variable in \code{df} .}

\item{covars}{A string specifying the covariates in \code{df} .}

\item{groupvar}{A string specifying the grouping variable in \code{df}.}

\item{ref}{Reset the ref level in risk_var}
}
\value{
A data frame combining group statistics (cases, total, incidence rate and Cox model results (HR, 95\% CI, beta, SE, p-value).
}
\description{
Extracts hazard ratios, coefficients, and summary statistics (cases, incidence rate)
from a Cox proportional hazards model and a data frame, for a specific group level.
}
