% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-cox_run_q.R
\name{cox_run_q}
\alias{cox_run_q}
\title{连续变量分位数Cox回归分析函数。}
\usage{
cox_run_q(
  data,
  mainvar,
  q = 3,
  time1 = NULL,
  time2 = NULL,
  timediff = NULL,
  event,
  covars = NULL,
  ...
)
}
\arguments{
\item{data}{数据框}

\item{mainvar}{连续变量名（字符串）}

\item{q}{分位数组数（整数，默认3）}

\item{time1}{随访开始时间变量名（字符串）}

\item{time2}{随访结束时间变量名（字符串）}

\item{timediff}{生存时间变量名（字符串，可选）}

\item{event}{结局变量名（字符串，0/1）}

\item{covars}{协变量名（字符串向量，可为NULL）}

\item{...}{传递给cox_run的其他参数}
}
\value{
一个列表，包含模型对象和分位数变量分析结果表
}
\description{
连续变量分位数Cox回归分析函数。
}
