% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-cox_run_sub.R
\name{cox_run_sub}
\alias{cox_run_sub}
\title{分组/亚组Cox回归批量分析函数，本函数支持基本的Cox回归分析和时间依赖（长数据）变量的Cox回归亚组分析。}
\usage{
cox_run_sub(
  data,
  group_var,
  time1 = NULL,
  time2 = NULL,
  timediff = NULL,
  event,
  mainvar,
  covars = NULL,
  extra_vars = NULL,
  ...
)
}
\arguments{
\item{data}{数据框}

\item{group_var}{分组变量名（字符串）}

\item{time1}{随访开始时间变量名（字符串）}

\item{time2}{随访结束时间变量名（字符串）}

\item{timediff}{生存时间变量名（字符串，可选）}

\item{event}{结局变量名（字符串，0/1）}

\item{mainvar}{关注变量名（字符串）}

\item{covars}{协变量名（字符串向量，可为NULL）}

\item{extra_vars}{额外关注变量名（字符串向量，可为NULL）}

\item{...}{传递给cox_run的其他参数}
}
\value{
一个数据框，包含分组、变量、HR、HR_lower、HR_upper、P、beta、se等
}
\description{
分组/亚组Cox回归批量分析函数，本函数支持基本的Cox回归分析和时间依赖（长数据）变量的Cox回归亚组分析。
}
